# specs/F04-content-administration/contracts/f04-openapi.yaml
openapi: 3.1.0
info:
  title: Pchinese F04 Content Administration API
  version: 1.0.0-design
  description: >
    Design contract for versioned ADMIN content management. Public read paths
    expose only publishable Free content; administrative mutations use the
    standard success/data/error/meta envelope and require expectedVersion.
    Management responses are typed safe projections, never persistence entities;
    lists are bounded and media intake is server-controlled.
servers:
  - url: /api/v1
...
paths:
...
  /admin/content/media:
    get:
      tags: [AdminContent]
      summary: List safe media-management projections for ADMIN
      operationId: listAdminMedia
      parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Size' } ]
      responses:
        '200': { $ref: '#/components/responses/ContentListSuccess' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
...
  /admin/content/media/{mediaAssetId}:
    get:
      tags: [AdminContent]
      summary: Reload one safe media projection for ADMIN
      operationId: getAdminMedia
      parameters: [ { $ref: '#/components/parameters/MediaAssetId' } ]
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /topics/{topicId}/publish:
    post:
      tags: [AdminContent]
      summary: Publish a topic
      operationId: publishTopic
      parameters: [ { $ref: '#/components/parameters/TopicId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
  /topics/{topicId}/unpublish:
    post:
      tags: [AdminContent]
      summary: Unpublish a topic
      operationId: unpublishTopic
      parameters: [ { $ref: '#/components/parameters/TopicId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
  /topics/{topicId}/archive:
    post:
      tags: [AdminContent]
      summary: Archive a topic
      operationId: archiveTopic
      parameters: [ { $ref: '#/components/parameters/TopicId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
...
  /lessons/{lessonId}/publish:
    post:
      tags: [AdminContent]
      summary: Publish a lesson
      operationId: publishLesson
      parameters: [ { $ref: '#/components/parameters/LessonId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
  /lessons/{lessonId}/unpublish:
    post:
      tags: [AdminContent]
      summary: Unpublish a lesson
      operationId: unpublishLesson
      parameters: [ { $ref: '#/components/parameters/LessonId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
  /lessons/{lessonId}/archive:
    post:
      tags: [AdminContent]
      summary: Archive a lesson
      operationId: archiveLesson
      parameters: [ { $ref: '#/components/parameters/LessonId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
...
  /segments/{segmentId}/publish:
    post:
      tags: [AdminContent]
      summary: Publish a segment
      operationId: publishSegment
      parameters: [ { $ref: '#/components/parameters/SegmentId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
  /segments/{segmentId}/unpublish:
    post:
      tags: [AdminContent]
      summary: Unpublish a segment
      operationId: unpublishSegment
      parameters: [ { $ref: '#/components/parameters/SegmentId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
  /segments/{segmentId}/archive:
    post:
      tags: [AdminContent]
      summary: Archive a segment
      operationId: archiveSegment
      parameters: [ { $ref: '#/components/parameters/SegmentId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
  /media:
    post:
      tags: [AdminContent]
      summary: Register one validated YouTube video reference for review
      operationId: createMedia
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MediaCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
  /media/{mediaAssetId}:
    patch:
      tags: [AdminContent]
      summary: Update YouTube media display metadata with optimistic versioning
      operationId: updateMedia
      parameters: [ { $ref: '#/components/parameters/MediaAssetId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MediaUpdateRequest' }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
  /media/{mediaAssetId}/approve:
    post:
      tags: [AdminContent]
      summary: Approve a clean pending-review media asset
      operationId: approveMedia
      parameters: [ { $ref: '#/components/parameters/MediaAssetId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
  /media/{mediaAssetId}/reject:
    post:
      tags: [AdminContent]
      summary: Reject a pending-review or approved media asset and withdraw dependent published lessons
      operationId: rejectMedia
      parameters: [ { $ref: '#/components/parameters/MediaAssetId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
  /media/{mediaAssetId}/quarantine:
    post:
      tags: [AdminContent]
      summary: Quarantine a pending-review or approved media asset and withdraw dependent published lessons
      operationId: quarantineMedia
      parameters: [ { $ref: '#/components/parameters/MediaAssetId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VersionedCommandRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    TopicId: { name: topicId, in: path, required: true, schema: { type: string, format: uuid } }
    LessonId: { name: lessonId, in: path, required: true, schema: { type: string, format: uuid } }
    SegmentId: { name: segmentId, in: path, required: true, schema: { type: string, format: uuid } }
    MediaAssetId: { name: mediaAssetId, in: path, required: true, schema: { type: string, format: uuid } }
    Page: { name: page, in: query, required: false, schema: { type: integer, minimum: 0, default: 0 } }
    Size: { name: size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  responses:
    ContentSuccess:
      description: Latest safe versioned content projection
      content: { application/json: { schema: { $ref: '#/components/schemas/ContentEnvelope' } } }
    ContentListSuccess:
      description: Paginated safe content-management projections
      content: { application/json: { schema: { $ref: '#/components/schemas/ContentListEnvelope' } } }
    ValidationError:
      description: Safe input or lifecycle validation error
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Unauthenticated:
      description: Missing, expired, revoked or invalid access credential
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    AuthorizationDenied:
      description: Caller lacks ADMIN role
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Resource absent or unavailable
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    StateConflict:
      description: Stale expectedVersion or forbidden state transition
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
...
    MediaCreateRequest:
      description: Register one YouTube video for review. The reference accepts a supported HTTPS YouTube watch, share, or embed link, or its canonical 11-character video ID. The backend stores only the canonical ID; it does not upload or download video bytes.
      type: object
      required: [youtubeVideoReference, title]
      properties:
        youtubeVideoReference: { type: string, minLength: 11, maxLength: 2048 }
        durationMilliseconds: { type: integer, minimum: 0 }
        title: { type: string, minLength: 1, maxLength: 200 }
        altText: { type: string, maxLength: 500 }
      additionalProperties: false
    MediaUpdateRequest:
      description: The YouTube source is immutable after creation. Register a distinct media asset to use a different video.
      type: object
      required: [expectedVersion]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        altText: { type: string, maxLength: 1000 }
        expectedVersion: { type: integer, minimum: 0 }
      additionalProperties: false
    VersionedCommandRequest:
      type: object
      required: [expectedVersion]
      properties: { expectedVersion: { type: integer, minimum: 0 } }
      additionalProperties: false
    ContentProjection:
      description: Safe typed ADMIN projection. It never includes persistence-only identifiers, actor identifiers, learner data, submitted provider URLs, storage keys, or verification diagnostics.
      type: object
      required: [id, entityType, publicationStatus, version]
      properties:
        id: { type: string, format: uuid }
        entityType: { type: string, enum: [TOPIC, LESSON, SEGMENT, MEDIA] }
        publicationStatus: { type: string, enum: [DRAFT, PUBLISHED, UNPUBLISHED, ARCHIVED, PENDING_SCAN, APPROVED, REJECTED, QUARANTINED] }
        version: { type: integer, minimum: 0 }
        title: { type: string, maxLength: 200 }
        parentId: { type: string, format: uuid }
        sequenceNo: { type: integer, minimum: 1 }
        accessLevel: { type: string, enum: [FREE] }
        providerName: { type: string, const: YOUTUBE }
        youtubeVideoId: { type: string, pattern: '^[A-Za-z0-9_-]{11}$' }
        slug: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        summary: { type: string, maxLength: 4000 }
        hskLevel: { type: integer, minimum: 1, maximum: 9 }
        transcriptHanzi: { type: string, maxLength: 20000 }
        transcriptPinyin: { type: string, maxLength: 20000 }
        translationVi: { type: string, maxLength: 20000 }
        dictationHint: { type: string, maxLength: 4000 }
        startMilliseconds: { type: integer, minimum: 0 }
        endMilliseconds: { type: integer, minimum: 0 }
        segmentType: { type: string, enum: [DICTATION, SHADOWING, BOTH, PLAYBACK_ONLY] }
        mediaKind: { type: string, const: VIDEO }
        altText: { type: string, maxLength: 500 }
        approvalEligible: { type: boolean, description: Safe indication that PENDING_SCAN YouTube media may be approved after trusted-source verification. }
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    Meta:
      type: object
      required: [correlationId]
      properties: { correlationId: { type: string, format: uuid } }
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, enum: [VALIDATION_ERROR, UNAUTHENTICATED, AUTHORIZATION_DENIED, RESOURCE_NOT_FOUND, STATE_CONFLICT] }
        message: { type: string }
    ContentEnvelope:
      type: object
      required: [success, data, error, meta]
      properties:
        success: { const: true }
        data: { $ref: '#/components/schemas/ContentProjection' }
        error: { type: 'null' }
        meta: { $ref: '#/components/schemas/Meta' }
    ContentListEnvelope:
      type: object
      required: [success, data, error, meta]
      properties:
        success: { const: true }
        data: { type: array, items: { $ref: '#/components/schemas/ContentProjection' } }
        error: { type: 'null' }
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
            - $ref: '#/components/schemas/PageMeta'
    ErrorEnvelope:
      type: object
      required: [success, data, error, meta]
      properties:
        success: { const: false }
        data: { type: 'null' }
        error: { $ref: '#/components/schemas/Error' }
        meta: { $ref: '#/components/schemas/Meta' }
security:
  - bearerAuth: []